requirements.txt
python-dotenv
fastapi
langchain
langsmith
langserve[all]
dask
paddlepaddle
paddleocr
openai-whisper
ffmpeg-python